+++ config.php
@@ -1,0 +1,4 @@
+<?php
+
+date_default_timezone_set('Asia/Almaty');
+
