+++ data.php
@@ -1,0 +1,55 @@
+<?php
+$user_name = 'Mansur';
+
+$categories  = [
+    'boards' => 'Доски и лыжи' ,
+    'attachment' => 'Крепления',
+    'boots' => 'Ботинки',
+    'clothing' => 'Одежда',
+    'tools' => 'Инструменты',
+    'other' => 'Разное'
+];
+$lots = [
+    [
+        'title' => '2014 Rossignol District Snowboard',
+        'type' => $categories['boards'],
+        'price' => 10999,
+        'image' => 'img/lot-1.jpg',
+        'timer' => '2021-01-29 16:16',
+    ],
+    [
+        'title' => 'DC Ply Mens 2016/2017 Snowboard',
+        'type' => $categories['boards'],
+        'price' => 159999,
+        'image' => 'img/lot-2.jpg',
+        'timer' => '2021-01-29 11:22',
+    ],
+    [
+        'title' => 'Крепления Union Contact Pro 2015 года размер L/XL',
+        'type' => $categories['attachment'],
+        'price' => 8000,
+        'image' => 'img/lot-3.jpg',
+        'timer' => '2021-01-28 19:40',
+    ],
+    [
+        'title' => 'Ботинки для сноуборда DC Mutiny Charocal',
+        'type' =>  $categories['boots'],
+        'price' => 10999,
+        'image' => 'img/lot-4.jpg',
+        'timer' => '2021-01-29 21:21',
+    ],
+    [
+        'title' => 'Куртка для сноуборда DC Mutiny Charocal',
+        'type' => $categories['clothing'],
+        'price' => 7500,
+        'image' => 'img/lot-5.jpg',
+        'timer' => '2021-01-29 23:23',
+    ],
+    [
+        'title' => 'Маска Oakley Canopy',
+        'type' => $categories['other'],
+        'price' => 5400,
+        'image' => 'img/lot-6.jpg',
+        'timer' => '2020-01-29 23:44',
+    ]
+];
